fixup!  gnrc/rpl/control_messages: `_ip_addr_str` fn

+++ sys/net/gnrc/routing/rpl/gnrc_rpl_control_messages.c
@@ -513,7 +513,7 @@ static inline uint32_t _sec_to_ms(uint32_t sec)
 
 static inline char *_ip_addr_str(ipv6_addr_t *addr)
 {
-    return ipv6_addr_to_str(addr_str, addr, (unsigned)sizeof(addr_str));
+    return ipv6_addr_to_str(addr_str, addr, sizeof(addr_str));
 }
 
 /** @todo allow target prefixes in target options to be of variable length */
